saveto cli command
