+++ dagreader/main.go
@@ -7,34 +7,19 @@ import (
 	"syscall"
 
 	"github.com/ethereum/go-ethereum/log"
-	"github.com/ethereum/go-ethereum/params"
 	"github.com/urfave/cli"
-
-	"github.com/Fantom-foundation/lachesis-dag-tool/dagreader/neo4j"
 )
 
 var (
-	// Git SHA1 commit hash of the release (set via linker flags).
-	gitCommit = ""
-	gitDate   = ""
 	// App that holds all commands and flags.
 	App = cli.NewApp()
 
-	neo4jUrlFlag = cli.StringFlag{
-		Name:  "neo4j",
-		Usage: "Neo4j DB url",
-		Value: neo4j.DefaultDb,
+	operaApiUrlFlag = cli.StringFlag{
+		Name:  "api",
+		Usage: "opera API url",
+		Value: "ws://127.0.0.1:4500",
 	}
 )
-
-// version of the current release
-func version() string {
-	params.VersionMajor = 0
-	params.VersionMinor = 1
-	params.VersionPatch = 0
-	params.VersionMeta = "rc.1"
-	return params.VersionWithCommit(gitCommit, gitDate)
-}
 
 func init() {
 	log.Root().SetHandler(
@@ -42,10 +27,10 @@ func init() {
 
 	App.Version = version()
 	App.Flags = []cli.Flag{
-		neo4jUrlFlag,
+		operaApiUrlFlag,
 	}
 	App.Commands = []cli.Command{
-		cmdListen,
+		cmdSaveTo,
 		cmdReadNeo4j,
 	}
 }
